Update user_model.php
Use built-in "is_unique" check for username/email on registration. The only downside of this is that the error message is less helpful: "The %s field must contain a unique value", but this is much more concise.

I believe that ```$CI->form_validation->set_message()``` could be used to provide a more helpful error message.